requirements.txt
Pillow
fpdf
pdf2image
pymupdf
